backend/requirements.txt
fastapi
uvicorn[standard]
pydantic>=2.7.0
langgraph
google-generativeai
python-multipart
python-dotenv
httpx
weaviate-client
